Extract idiomatic test-fixtures plugin to buildSrc
Plugin is applied to all "groovy projects", see `groovyProject.gradle`.

The plugin configures the Project as a test fixtures producer if
`src/testFixtures` is a directory.

The plugin configures the Project as a test fixtures consumer according
to the `testFixtures` extension configuration.

No more Groovy Closure with optional parameters as a Project extra
property.

Signed-off-by: Paul Merlin <paul@gradle.com>